data_streaming_kafka/requirements.txt
Flask
pymysql
kafka-python
requests
docker
